moving types to EE file
