+++ app/client/src/ce/constants/ModuleInstanceConstants.ts
@@ -1,5 +1,4 @@
 import type { MODULE_TYPE } from "@appsmith/constants/ModuleConstants";
-import type { ActionResponse } from "api/ActionAPI";
 
 export type ModuleId = string;
 export type ModuleInstanceId = string;
@@ -17,10 +16,3 @@ export interface ModuleInstance {
   };
   sourceModuleId: ModuleId;
 }
-
-export interface ModuleInstanceData {
-  config: ModuleInstance;
-  data: ActionResponse | undefined;
-  isLoading: boolean;
-}
-export type ModuleInstanceDataState = ModuleInstanceData[];
